% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{B2_pri}
\alias{B2_pri}
\title{Function to set a B2 prior}
\usage{
B2_pri(b, p, q)
}
\arguments{
\item{b}{scale parameter.}

\item{p}{shape parameter.}

\item{q}{shape parameter.}
}
\value{
A list with the prior name and all the quantities required to implement it within the sampler.
}
\description{
Allow the user to specify a B2 prior for a variance parameter
}
