% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_summaries.R
\name{burn_thin}
\alias{burn_thin}
\title{Burn-in and thinning of posterior drawns}
\usage{
burn_thin(out, n_burn, n_thin = 1)
}
\arguments{
\item{out}{Object created by \code{sample_modelDSD} or \code{sample_model_manual}.}

\item{n_burn}{Length burn-in period.}

\item{n_thin}{Length thinning interval.}
}
\description{
Providing as input an object created by \code{sample_modelDSD} and \code{sample_model_manual}, the function performs burn-in and thinning of the chain.
}
