% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{DSD_pri}
\alias{DSD_pri}
\title{Function to set a DSD prior in a model}
\usage{
DSD_pri(alpha, beta, alpha_til, beta_til, b, p, q)
}
\arguments{
\item{alpha}{parameter related to the design.}

\item{beta}{parameter related to the design.}

\item{alpha_til}{parameter related to the precision matrix.}

\item{beta_til}{parameter related to the precision matrix.}

\item{b}{scale parameter of the underlying B2 prior.}

\item{p}{shape parameter of the underlying B2 prior.}

\item{q}{shape parameter of the underlying B2 prior.}
}
\value{
A list with the prior name and all the quantities required to implement it within the sampler.
}
\description{
Allow the user to specify a DSD prior for the scale parameter related to a Gaussian random effect
}
