% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{DSD_pri_beta}
\alias{DSD_pri_beta}
\title{Function to set a DSD prior on the scale parameter of beta}
\usage{
DSD_pri_beta(alpha, x, b, q)
}
\arguments{
\item{alpha}{parameter related to the design.}

\item{x}{covariate vector.}

\item{b}{scale parameter of the undelying B2 prior.}

\item{q}{shape parameter of the undelying B2 prior.}
}
\value{
A list with the prior name and all the quantities required to implment it within the sampler.
}
\description{
Allow the user to specify a DSD prior for the scale parameter for the regression coefficient beta prior.
}
