% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{Gaussian_pri_beta}
\alias{Gaussian_pri_beta}
\title{Function to set a standard prior on beta}
\usage{
Gaussian_pri_beta()
}
\value{
A list with the prior name and all the quantities required to implement it within the sampler.
}
\description{
Allow the user to specify a standard Gaussian prior for the regression coefficient beta.
}
