% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{get_b_DSD}
\alias{get_b_DSD}
\title{Prior elicitation under B2 prior}
\usage{
get_b_DSD(pi_0 = 0.5, c, n, p_DSD, q_DSD, n_sim = 1e+06)
}
\arguments{
\item{c}{fixed quantile.}

\item{n}{data sample size.}

\item{p_DSD}{value of the parameter \code{p} of the DSD prior.}

\item{q_DSD}{value of the parameter \code{q} of the DSD prior.}

\item{n_sim}{number of Monte Carlo replicates to evaluate the quantile.}

\item{pi}{probability related to the quantile.}
}
\value{
The parameter \code{b} that delivers a prior distribution for the random effect sample variance \eqn{V} satisfying the relation \deqn{
P[V>c]=pi
}
}
\description{
Allow to retrieve the B2 scale parameter b under the considered probabilistic statement
}
