% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{get_DSD_par}
\alias{get_DSD_par}
\title{Computing useful quantities for DSD prior}
\usage{
get_DSD_par(reff)
}
\arguments{
\item{reff}{An object of class \code{"reff"} produced by the function \code{\link{get_reff}}}
}
\value{
A list with \eqn{\tilde{\alpha}}, \eqn{\tilde{\beta}} and the mean of the non-null eigenvalues \eqn{\lambda}.
}
\description{
Starting from a list produced by the function \code{\link{get_reff}}, the quantities required to implement a DSD prior are computed.
}
