% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{get_reff}
\alias{get_reff}
\title{Computing useful quantities about random effects}
\usage{
get_reff(Z, K, K_full_rank = FALSE)
}
\arguments{
\item{Z}{Design matrix.}

\item{K}{Precision matrix.}

\item{K_full_rank}{Logical. If TRUE the function assumes that \code{K} is full-rank}
}
\value{
A list with \code{Z}, \code{K}, the rank of \code{K}, the number of columns of \code{K}, the matrix A that contain a basis
of the null space, the vector e with the vector employed to set the linear constraint and \code{K_full_rank}.
}
\description{
Starting from the design matrix and the precision matrix, some details about the random effect are returned.
}
