% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{graph_munich}
\alias{graph_munich}
\title{Spatial graph of Munich data}
\format{
An object of class \code{inla.graph} of length 4.
}
\usage{
graph_munich
}
\description{
Spatial graph of Munich data
}
\keyword{datasets}
