% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_summaries.R
\name{posterior_summaries}
\alias{posterior_summaries}
\title{Computation of posterior summaries}
\usage{
posterior_summaries(model_out)
}
\arguments{
\item{model_out}{Object created by \link{\code{sample_modelDSD}}, \link{\code{sample_model_manual}} or \link{\code{burn_thin}}.}
}
\description{
Providing as input an object created by \link{\code{sample_modelDSD}}, \link{\code{sample_model_manual}} or \link{\code{burn_thin}}, the function computes posterior summaries.
}
