% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_model.R
\name{sample_model_manual}
\alias{sample_model_manual}
\title{Draw posterior samples from a Bayesian linear mixed model}
\usage{
sample_model_manual(
  y,
  X,
  reff_list,
  model = c("Gaussian", "Bernoulli_logit", "Bernoulli_probit"),
  pri_s2e = NULL,
  pri_s2g,
  pri_s2b,
  beta_init = rep(0, ncol(X)),
  s2g_init = rep(1, length(reff_list)),
  s2e_init = 1,
  niter = 25000,
  pr = niter/10,
  thin = 10
)
}
\arguments{
\item{y}{Vector of observed responses.}

\item{X}{Matrix with fixed effects. The intercept is added by the function.}

\item{reff_list}{list with the design matrices of the random effects.}

\item{model}{Model to be fitted on input data. Available 'Gaussian', 'Bernoulli_logit' and 'Bernoulli_probit'.}

\item{pri_s2e}{Prior for the variance parameter in case of Gaussian model.}

\item{pri_s2g}{List of priors for the scale hyperparameters ruling the random effects.}

\item{pri_s2b}{List of priors for the regression coefficients related to fixed effects.}

\item{beta_init}{Initial values for the vector of regression coefficients. Default all 0s.}

\item{s2g_init}{Vector with the initial values for the random effects scales. Default all 1s.}

\item{s2e_init}{Initial value for the data variance. Default 1.}

\item{niter}{Number of MCMC iterations.}

\item{pr}{Frequency of iterations for printing the progress.}

\item{thin}{Number of thinned observations.}
}
\value{
The function returns a list with the drawns from the posterior ditribution of the model parameters.
}
\description{
Allow the user to specify a Bayesian linear mixed model, with flexibility in the prior choices.
}
