% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_model.R
\name{sample_modelDSD}
\alias{sample_modelDSD}
\title{Draw posterior samples from a Bayesian linear mixed model}
\usage{
sample_modelDSD(
  y,
  X,
  reff_list,
  pi_0 = 0.5,
  p_DSD = 0.5,
  q_DSD = 1.5,
  model = c("Gaussian", "Bernoulli_logit", "Bernoulli_probit"),
  beta_init = rep(0, ncol(X)),
  s2g_init = rep(1, length(reff_list)),
  s2e_init = 1,
  niter = 25000,
  pr = niter/10,
  thin = 10
)
}
\arguments{
\item{y}{Vector of observed responses.}

\item{X}{Matrix with fixed effects. The intercept is added by the function.}

\item{reff_list}{list with the design matrices of the random effects.}

\item{pi_0}{Prior elicitation parameter.}

\item{p_DSD}{Hyperparameter determining the prior behavior near by 0. Default 0.5.}

\item{q_DSD}{Hyperparameter determining the tail decay of the prior. Default 1.5.}

\item{model}{Model to be fitted on input data. Available 'Gaussian', 'Bernoulli_logit' and 'Bernoulli_probit'.}

\item{beta_init}{Initial values for the vector of regression coefficients. Default all 0s.}

\item{s2g_init}{Vector with the initial values for the random effects scales. Default all 1s.}

\item{s2e_init}{Initial value for the data variance. Default 1.}

\item{niter}{Number of MCMC iterations.}

\item{pr}{Frequency of iterations for printing the progress.}

\item{thin}{Number of thinned observations.}
}
\value{
The function returns a list with the drowns from the posterior distribution of the model parameters.
}
\description{
Allow the user to specify a Bayesian linear mixed model, with flexibility in the prior choices.
}
