% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tokyo}
\alias{tokyo}
\title{Tokyo rainfall data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 731 rows and 2 columns.
}
\usage{
tokyo
}
\description{
Tokyo rainfall data
}
\keyword{datasets}
