% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{W_pri}
\alias{W_pri}
\title{Function to set a Weibull prior}
\usage{
W_pri(lambda, kappa)
}
\arguments{
\item{lambda}{shape parameter.}

\item{kappa}{scale parameter.}
}
\value{
A list with the prior name and all the quantities required to implement it within the sampler.
}
\description{
Allow the user to specify a Weibull prior for a variance parameter
}
