% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_summaries.R
\name{compute_linpred}
\alias{compute_linpred}
\title{Evaluate linear predictor}
\usage{
compute_linpred(out, X, Z_list)
}
\arguments{
\item{out}{Object created by \link{\code{sample_modelDSD}}, \link{\code{sample_model_manual}} or \link{\code{burn_thin}}.}

\item{X}{Fixed effects design matrix.}

\item{Z_list}{List of design matrices related to random effects.}
}
\description{
Receiving as input the result of \link{\code{sample_modelDSD}}, \link{\code{sample_model_manual}} or \link{\code{burn_thin}} functions, computes the linear predictor in the points provided throug the design matrices.
}
